+++ cite_seq_count/__main__.py
@@ -1,29 +1,24 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python3.6
 """
 Author: Patrick Roelli
 """
-import csv
-import gzip
 import sys
 import time
 import os
 import datetime
-import shutil
+import pkg_resources
 
 from argparse import ArgumentParser
 from argparse import RawTextHelpFormatter
 from collections import OrderedDict
-from itertools import islice, combinations
+from collections import Counter
 
-import Levenshtein
-from multiprocess import Process, cpu_count, Pool
-import pkg_resources
-import regex
-from scipy import sparse
-from scipy.io import mmwrite
-from numpy import int16
+from multiprocess import cpu_count
+from multiprocess import Pool
 
+from cite_seq_count import preprocessing
 from cite_seq_count import processing
+from cite_seq_count import io
 from cite_seq_count import secondsToText
 
 version = pkg_resources.require("cite_seq_count")[0].version
@@ -106,6 +101,15 @@ def get_args():
         required=False, type=int, default=2,
         help=("Maximum hamming distance allowed for antibody barcode.")
     )
+    allowed_errors = ['i','s','d','e','i,s','i,d','s,d']
+    filters.add_argument(
+        '-e', '--error-type', dest='error_type',
+        required=False, type=str, default='s',
+        help=("Error type for the regex match."
+            "\ni: Insertions only\ns: Substitutions only\n"
+            "d: Deletion only\ne: All of the above"),
+        choices=allowed_errors
+    )
     
     # Remaining arguments.
     parser.add_argument('-T', '--threads', required=False, type=int,
@@ -152,327 +156,6 @@ def get_args():
     return parser
 
 
-def parse_whitelist_csv(filename, barcode_length):
-    """Reads white-listed barcodes from a CSV file.
-
-    The function accepts plain barcodes or even 10X style barcodes with the
-    `-1` at the end of each barcode.
-
-    Args:
-        filename (str): Whitelist barcode file.
-        barcode_length (int): Length of the expected barcodes.
-
-    Returns:
-        set: The set of white-listed barcodes.
-
-    """
-    STRIP_CHARS = '0123456789- \t\n'
-    with open(filename, mode='r') as csv_file:
-        csv_reader = csv.reader(csv_file)
-        whitelist = [row[0].strip(STRIP_CHARS) for row in csv_reader
-                     if (len(row[0].strip(STRIP_CHARS)) == barcode_length)]
-    return set(whitelist)
-
-
-def parse_tags_csv(filename):
-    """Reads the TAGs from a CSV file.
-
-    The expected file format (no header) is: TAG,TAG_NAME.
-    e.g. file content
-        GTCAACTCTTTAGCG,Hashtag_1
-        TGATGGCCTATTGGG,Hashtag_2
-        TTCCGCCTCTCTTTG,Hashtag_3
-
-    Args:
-        filename (str): TAGs file.
-
-    Returns:
-        dict: A dictionary containing the TAGs and their names.
-
-    """
-    with open(filename, mode='r') as csv_file:
-        csv_reader = csv.reader(csv_file)
-        tags = {}
-        for row in csv_reader:
-            tags[row[0].strip()] = row[1].strip()
-    return tags
-
-
-def check_tags(tags, maximum_distance):
-    """Evaluates the distance between the TAGs based on the `maximum distance`
-    argument provided.
-
-    Additionally, it adds the barcode to the name of the TAG circumventing the
-    need of having to share the mapping of the antibody and the barcode.
-    
-    The output will have the keys sorted by TAG length (longer first). This
-    way, longer barcodes will be evaluated first.
-
-    Args:
-        tags (dict): A dictionary with the TAGs + TAG Names.
-        maximum_distance (int): The maximum Levenshtein distance allowed
-            between two TAGs.
-
-    Returns:
-        collections.OrderedDict: An ordered dictionary containing the TAGs and
-            their names in descendent order based on the length of the TAGs.
-
-    """
-    ordered_tags = OrderedDict()
-    for tag in sorted(tags, key=len, reverse=True):
-        ordered_tags[tag] = tags[tag] + '-' + tag
-
-    # If only one TAG is provided, then no distances to compare.
-    if (len(tags) == 1):
-        return(ordered_tags)
-    
-    offending_pairs = []
-    for a, b in combinations(ordered_tags.keys(), 2):
-        distance = Levenshtein.distance(a, b)
-        if (distance <= maximum_distance):
-            offending_pairs.append([a, b, distance])
-    
-    # If offending pairs are found, print them all.
-    if offending_pairs:
-        print(
-            '[ERROR] Minimum Levenshtein distance of TAGs barcode is less '
-            'than given threshold.\n'
-            'Please use a smaller distance.\n\n'
-            'Offending case(s):\n'
-        )
-        for pair in offending_pairs:
-            print(
-                '\t{tag1}\n\t{tag2}\n\tDistance = {distance}\n'
-                .format(tag1=pair[0], tag2=pair[1], distance=pair[2])
-            )
-        sys.exit('Exiting the application.\n')
-    
-    return(ordered_tags)
-
-
-def get_read_length(filename):
-    """Check wether SEQUENCE lengths are consistent in a FASTQ file and return
-    the length.
-
-    Args:
-        filename (str): FASTQ file.
-
-    Returns:
-        int: The file's SEQUENCE length.
-
-    """
-    with gzip.open(filename, 'r') as fastq_file:
-        secondlines = islice(fastq_file, 1, 1000, 4)
-        temp_length = len(next(secondlines).rstrip())
-        for sequence in secondlines:
-            read_length = len(sequence.rstrip())
-            if (temp_length != read_length):
-                sys.exit(
-                    '[ERROR] Sequence length is not consistent. Please, trim all '
-                    'sequences at the same length.\n'
-                    'Exiting the application.\n'
-                )
-    return(read_length)
-
-
-def check_read_lengths(read1_length, cb_first, cb_last, umi_first, umi_last):
-    """Check Read1 length against CELL and UMI barcodes length.
-
-    Args:
-        read1_length (int): Read1 length.
-        cb_first (int): Barcode first base position for Read1.
-        cb_last (int): Barcode last base position for Read1.
-        umi_first (int): UMI first base position for Read1.
-        umi_last (int): UMI last base position for Read1.
-
-    Returns:
-        slice: A `slice` object to extract the Barcode from the sequence string.
-        slice: A `slice` object to extract the UMI from the sequence string.
-        int: The Barcode + UMI length.
-
-    """
-    barcode_length = cb_last - cb_first + 1
-    umi_length = umi_last - umi_first + 1
-    barcode_umi_length = barcode_length + umi_length
-    barcode_slice = slice(cb_first - 1, cb_last)
-    umi_slice = slice(umi_first - 1, umi_last)
-
-    if barcode_umi_length > read1_length:
-        sys.exit(
-            '[ERROR] Read1 length is shorter than the option you are using for '
-            'Cell and UMI barcodes length. Please, check your options and rerun.\n\n'
-            'Exiting the application.\n'
-        )
-    elif barcode_umi_length < read1_length:
-        print(
-            '[WARNING] Read1 length is {}bp but you are using {}bp for Cell '
-            'and UMI barcodes combined.\nThis might lead to wrong cell '
-            'attribution and skewed umi counts.\n'
-            .format(read1_length, barcode_umi_length)
-        )
-    
-    return(barcode_slice, umi_slice, barcode_umi_length)
-
-
-def generate_regex(tags, maximum_distance, legacy=False, max_poly_a=6, read2_length=98, user_regex=None):
-    """Generate regex based ont he provided TAGs.
-
-    Args:
-        tags (dict): A dictionary with the TAGs + TAG Names.
-        maximum_distance (int): The maximum Levenshtein distance allowed
-            between two TAGs.
-        legacy (bool): `True` if you use an earlier version of the kit that adds
-            a T, C, or G at the end and you expect polyA tails in the data.
-            Default is False.
-        max_poly_a (int): Run length of A's expected for the polyA tail. Default
-            is 6.
-        read2_length (int): Length of Read2. Default is 98.
-        user_regex (str): A regular expression to use for TAG matching. Default
-            is None.
-
-    Returns:
-        regex.Pattern: An object that matches against any of the provided TAGs
-            within the maximum distance provided.
-
-    """
-    # Get a list of the available TAGs.
-    tag_keys = tags.keys()
-
-    # Get the length of the longest TAG.
-    longest_ab_tag = len(next(iter(tags)))
-
-    if user_regex:
-        # If more than one TAG is provided and their length is different, issue a
-        # warning.
-        if len(tag_keys) > 1:
-            for i in range(1, len(tag_keys)):
-                if len(tag_keys[i]) != len(tag_keys[i - 1]):
-                    print(
-                        '[WARNING] Different length TAGs have been provided while '
-                        'you specified a custom Regex. An OR method is recommended '
-                        'for this scenarios. No additional validations will be '
-                        'applied. Use it at your own risk.\n'
-                    )
-                    break
-        
-        regex_pattern = regex.compile(user_regex)
-        return(regex_pattern)
-
-    elif legacy:
-        # Keep the minimum value between `max_poly_a` provided and the remaining
-        # length of the read after removing the barcode length.
-        polya_run = min(max_poly_a, read2_length - longest_ab_tag - 1)
-
-        # Read comment below for `e` meaning in the regex.
-        pattern = r'(^(\L<options>)[TGC][A]{{{},}}){{s<={}}}'.format(
-            polya_run, maximum_distance
-        )
-
-    else:
-        # `e` is part of the `regex` fuzzy logic: it means `error` in general,
-        # whether it's a (s)ubstitution, (i)nsertion or (d)eletion. In this 
-        # case, it means it allows `maximum_distance` errors to happen.
-        pattern = r'(^\L<options>){{s<={}}}'.format(maximum_distance)
-
-    # Compiling the regex makes it run faster.
-    regex_pattern = regex.compile(pattern, options=tag_keys)
-    return(regex_pattern)
-
-
-
-
-def write_to_files(sparse_matrix, final_results, ordered_tags_map, data_type, outfile):
-    """Write the sparse matrices to files in mtx format.
-
-    Args:
-        sparse_matrix (dok_matrix): Results in a sparse matrix.
-        final_results (dict): Results in a dict of dicts of Counters.
-        ordered_tags_map (dict): Tags in order with indexes as values.
-        data_type (string): A string definning if the data is umi or read based.
-        oufile (string): Path to the mtx file.
-    
-    """
-    prefix = data_type + '_count'
-    os.makedirs(prefix, exist_ok=True)
-    mmwrite(os.path.join(prefix,outfile),sparse_matrix)
-    with gzip.open(os.path.join(prefix,'barcodes.tsv.gz'), 'wb') as barcode_file:
-        for barcode in final_results:
-            barcode_file.write('{}\n'.format(barcode).encode())
-    with gzip.open(os.path.join(prefix,'features.tsv.gz'), 'wb') as feature_file:
-        for feature in ordered_tags_map:
-            feature_file.write('{}\n'.format(feature).encode())
-    with open(os.path.join(prefix,outfile),'rb') as mtx_in:
-        with gzip.open(os.path.join(prefix,outfile) + '.gz','wb') as mtx_gz:
-            shutil.copyfileobj(mtx_in, mtx_gz)
-    os.remove(os.path.join(prefix,outfile))
-
-def generate_sparse_matrices(final_results, ordered_tags_map, top_cells):
-    """
-    Create two sparse matrices with umi and read counts.
-
-    Args:
-        final_results (dict): Results in a dict of dicts of Counters.
-        ordered_tags_map (dict): Tags in order with indexes as values.
-
-    Returns:
-        umi_results_matrix (scipy.sparse.dok_matrix): UMI counts
-        read_results_matrix (scipy.sparse.dok_matrix): Read counts
-
-    """
-    umi_results_matrix = sparse.dok_matrix((len(ordered_tags_map) ,len(top_cells)), dtype=int16)
-    read_results_matrix = sparse.dok_matrix((len(ordered_tags_map) ,len(top_cells)), dtype=int16)
-    
-    for i,cell_barcode in enumerate(top_cells):
-        for j,TAG in enumerate(final_results[cell_barcode]):
-            if final_results[cell_barcode][TAG]:
-                umi_results_matrix[ordered_tags_map[TAG],i]=len(final_results[cell_barcode][TAG])
-                read_results_matrix[ordered_tags_map[TAG],i]=sum(final_results[cell_barcode][TAG].values())
-    return(umi_results_matrix, read_results_matrix)
-
-
-def blocks(files, size=65536):
-    """
-    A fast way of counting the lines of a large file.
-    Ref:
-        https://stackoverflow.com/a/9631635/9178565
-
-    Args:
-        files (io.handler): A file handler 
-        size (int): Block size
-    Returns:
-        A generator
-    """
-    while True:
-        b = files.read(size)
-        if not b: break
-        yield b
-
-
-def get_n_lines(file_path, top_n):
-    """
-    Determines how many lines have to be processed
-    depending on options and number of available lines.
-    Checks that the number of lines is a multiple of 4.
-
-    Args:
-        file_path (string): Path to a fastq.gz file
-        top_n (int): Number of reads to be used as defined by the user.
-
-    Returns:
-        n_lines (int): Number of lines to be used
-    """
-    print('Counting number of reads')
-    with gzip.open(file_path, "rt",encoding="utf-8",errors='ignore') as f:
-        n_lines = sum(bl.count("\n") for bl in blocks(f))
-    if n_lines %4 !=0:
-        sys.exit('{}\'s number of lines is not a multiple of 4. The file might be corrupted.\n Exiting')
-    if top_n and top_n*4 < n_lines:
-        n_lines = top_n*4
-    elif top_n and top_n*4 > n_lines:
-        print('The data only contains {} reads. Processing all the reads'.format(n_lines/4))
-    return(n_lines)
-
-
 def create_report(n_reads, reads_per_cell, no_match, version, start_time, ordered_tags_map, args):
     """
     Creates a report with details about the run in a yaml format.
@@ -500,8 +183,8 @@ reads processed: {}
 Percentage mapped: {}
 Percentage unmapped: {}
 Parameters:
-\tRead1: {}
-\tRead2: {}
+\tRead1_filename: {}
+\tRead2_filename: {}
 \tCell barcode:
 \t\tFirst position: {}
 \t\tLast position: {}
@@ -536,40 +219,36 @@ def main():
     # Parse arguments.
     args = parser.parse_args()
     if args.whitelist:
-        whitelist = parse_whitelist_csv(args.whitelist,
+        whitelist = preprocessing.parse_whitelist_csv(args.whitelist,
                                         args.cb_last - args.cb_first + 1)
     else:
         whitelist = None
 
     # Load TAGs/ABs.
-    ab_map = parse_tags_csv(args.tags)
-    ab_map = check_tags(ab_map, args.hamming_thresh)
+    ab_map = preprocessing.parse_tags_csv(args.tags)
+    ab_map = preprocessing.check_tags(ab_map, args.hamming_thresh)
     # Get reads length. So far, there is no validation for Read2.
-    read1_length = get_read_length(args.read1_path)
-    #read2_length = get_read_length(args.read2_path)
+    read1_length = preprocessing.get_read_length(args.read1_path)
+    read2_length = preprocessing.get_read_length(args.read2_path)
 
     # Check Read1 length against CELL and UMI barcodes length.
     (barcode_slice, 
      umi_slice, 
-     barcode_umi_length) = check_read_lengths(read1_length, args.cb_first,
+     barcode_umi_length) = preprocessing.check_read_lengths(read1_length, args.cb_first,
                                               args.cb_last, 
                                               args.umi_first, args.umi_last)
     
-    # Get unique combinations of Barcode+UMI+R2.
-    # unique_lines = get_unique_lines(
-    #     args.read1_path, args.read2_path, barcode_slice, umi_slice,
-    #     barcode_umi_length, args.first_n)
-
     # Generate the compiled regex pattern.
-    regex_pattern = generate_regex(ab_map, args.hamming_thresh, max_poly_a=6)
+    regex_pattern = preprocessing.generate_regex(ab_map, args.hamming_thresh, args.error_type, max_poly_a=6)
 
     
-    n_lines = get_n_lines(args.read1_path, args.first_n)  
+    n_lines = preprocessing.get_n_lines(args.read1_path, args.first_n)  
     
     n_reads = n_lines/4
     n_threads = args.n_threads
-    #Initiate multiprocessing leaving one thread available.
+    
     print('Started mapping')
+    #Run with one process
     if n_threads <= 1:
         print('CITE-seq-Count is running with only one core.')
         (final_results, merged_no_match) = processing.classify_reads_multi_process(
@@ -585,14 +264,22 @@ def main():
                 args.legacy,
                 args.debug)
         print('Mapping done')
+        umis_per_cell = Counter()
+        reads_per_cell = Counter()
+        for cell_barcode,counts in final_results.items():
+            umis_per_cell[cell_barcode] = sum([len(counts[UMI]) for UMI in counts if UMI != 'no_match' and UMI != 'bad_construct'])
+            reads_per_cell[cell_barcode] = sum([sum(counts[UMI].values()) for UMI in counts if UMI != 'no_match' and UMI != 'bad_construct'])
     else:
+        # Run with multiple processes
         p = Pool(processes=n_threads)
-        # We need the chunk size to be a multiple of 4 to always start of the right line in the fastq file
+        # We need the chunk size to be a multiple of 4 to always start on the correct line in the fastq file
         chunk_size=round((n_lines-1)/n_threads)
         while chunk_size % 4 != 0:
             chunk_size +=1
         chunks = range(1,n_lines,chunk_size)
+
         parallel_results = []
+
         for first_line in list(chunks):
            p.apply_async(processing.classify_reads_multi_process,
                 args=(
@@ -612,23 +299,22 @@ def main():
         p.close()
         p.join()
         print('Mapping done')
-
-
         print('Merging results')
-        (final_results,reads_per_cell, merged_no_match) = processing.merge_results(parallel_results)
+        (final_results, umis_per_cell, reads_per_cell, merged_no_match) = processing.merge_results(parallel_results)                
         del(parallel_results)
     
     ordered_tags_map = OrderedDict()
     for i,tag in enumerate(ab_map.values()):
         ordered_tags_map[tag] = i
     ordered_tags_map['no_match'] = i + 1
+    ordered_tags_map['bad_construct'] = i + 2
 
-    # if not whitelist:
-    top_cells_tuple = reads_per_cell.most_common(args.cells)
+
+    # Sort cells by number of mapped umis
+    top_cells_tuple = umis_per_cell.most_common(args.cells)
     top_cells = [pair[0] for pair in top_cells_tuple]
-    #     final_results = {key: final_results[key] for key in top_cells}
-
-    #We just need to add the missing cell barcodes.
+    
+    # Add potential missing cell barcodes.
     if whitelist:
         for missing_cell in args.whitelist:
             if missing_cell in final_results:
@@ -639,10 +325,10 @@ def main():
                     final_results[missing_cell][TAG] = 0
                 top_cells.append(missing_cell)
 
-    (umi_results_matrix, read_results_matrix) = generate_sparse_matrices(final_results, ordered_tags_map, top_cells)
+    (umi_results_matrix, read_results_matrix) = processing.generate_sparse_matrices(final_results, ordered_tags_map, top_cells)
     
-    write_to_files(umi_results_matrix, final_results, ordered_tags_map, 'umi', args.outfile)
-    write_to_files(read_results_matrix, final_results, ordered_tags_map, 'read', args.outfile)
+    io.write_to_files(umi_results_matrix, final_results, ordered_tags_map, 'umi', args.outfile)
+    io.write_to_files(read_results_matrix, final_results, ordered_tags_map, 'read', args.outfile)
       
     # Save no_match TAGs to `args.unknowns_file` file.
     if args.unknowns_file: